requirements.txt
# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0

# Google Cloud & AI
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
google-cloud-speech>=2.21.0
google-cloud-texttospeech>=2.14.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Async support
aiohttp>=3.9.0

# Web framework (for Twilio webhooks)
flask>=3.0.0
flask-cors>=4.0.0

# Telephony
twilio>=8.10.0

# Audio processing
pydub>=0.25.1
SpeechRecognition>=3.10.0

# Utilities
click>=8.1.0  # CLI framework
rich>=13.0.0  # Terminal formatting
tenacity>=8.2.0  # Retry logic
python-dateutil>=2.8.2  # Date parsing for auth flow

# Alternative TTS (simpler option)
gTTS>=2.4.0

# OpenAI (optional - for Whisper ASR)
openai>=1.3.0

# Healthcare-specific (placeholder - add as needed)
# fhir.resources>=7.0.0  # FHIR data models
# hl7apy>=1.3.0  # HL7 message parsing

# Development
ipython>=8.12.0
